move init tests to a separate file
