+++ tests/test_iotedgedev_solution.py
@@ -3,7 +3,6 @@ import os
 import platform
 import shutil
 import time
-import uuid
 from unittest import mock
 
 import pytest
@@ -109,35 +108,6 @@ def test_solution_create_in_empty_child_dir(prepare_solution_with_env):
     result = runner_invoke(['solution', 'new', dirname])
 
     assert 'AZURE IOT EDGE SOLUTION CREATED' in result.output
-
-
-def test_solution_init_without_name():
-    result = runner_invoke(['solution', 'init'], True)
-
-    assert "Directory is not empty" in result.output
-
-
-def test_solution_init_with_invalid_name_non_empty_dir():
-    dirname = f'test-{uuid.uuid4()}'
-    os.makedirs(f'{dirname}/empty_dir')
-
-    result = runner_invoke(['solution', 'init', dirname], True)
-
-    assert "Directory is not empty" in result.output
-    shutil.rmtree(dirname, ignore_errors=True)
-
-
-def test_solution_init_with_valid_name():
-    dirname = f'test-{uuid.uuid4()}'
-
-    # Mock calls to additional commands, to avoid triggering user prompts
-    with mock.patch('iotedgedev.utility.Utility.call_proc') as mock_call_proc:
-        result = runner_invoke(['solution', 'init', dirname], True)
-
-    assert 'AZURE IOT EDGE SOLUTION CREATED' in result.output
-    mock_call_proc.assert_called_with(["iotedgedev", "iothub", "setup", "--update-dotenv"])
-    assert mock_call_proc.call_count == 2
-    shutil.rmtree(dirname, ignore_errors=True)
 
 
 def test_solution_create_valid_runtime_tag():
